requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-mail==1.4.2
Flask==3.1.0
flask-swagger-ui==4.11.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.10.1
pydantic-mongo==2.3.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sendgrid==6.11.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
